1. Added test_dram_load_store cuda regression test to purely check the DRAM communication. 2. Added test_dram_device_allocated and test_dram_host_allocated cuda regression tests. 3. Removed deprecated test_dram cuda regression test.
