+++ cl_manycore/regression/cuda/Makefile.tests
@@ -11,7 +11,9 @@ INDEPENDENT_MAIN_TESTS += test_memset
 INDEPENDENT_MAIN_TESTS += test_barrier
 INDEPENDENT_MAIN_TESTS += test_packet 
 INDEPENDENT_MAIN_TESTS += test_stack_load
-INDEPENDENT_MAIN_TESTS += test_dram
+INDEPENDENT_MAIN_TESTS += test_dram_load_store
+INDEPENDENT_MAIN_TESTS += test_dram_host_allocated
+INDEPENDENT_MAIN_TESTS += test_dram_device_allocated
 INDEPENDENT_MAIN_TESTS += test_vec_add
 INDEPENDENT_MAIN_TESTS += test_vec_add_parallel
 INDEPENDENT_MAIN_TESTS += test_vec_add_parallel_multi_grid
