+++ cl_manycore/regression/cuda/test_dram_device_allocated.h
@@ -1,0 +1,15 @@
+#ifndef TEST_DRAM_DEVICE_ALLOCATED_H
+#define TEST_DRAM_DEVICE_ALLOCATED_H
+
+
+#include <stdio.h>
+#include <sys/ioctl.h>
+#include <unistd.h>
+#include <string.h>
+#include <time.h>
+#include <stdlib.h>
+
+#include "cuda_tests.h"
+
+
+#endif
